Struct binartree::tree::BinaryTree

pub struct BinaryTree<T> where
    T: Copy + Clone + Ord + Eq,  { /* fields omitted */ }

Realisation of Binary Search Tree in Rust lang

English: ABOUT STRUCTURE

Binary Search Tree is a directed graph, each node of it has a key - a certain value that can be compared and two branches - right and left. (we will assume that tree always have branches, but they can be empty, or contain a value).

The left branch stores nodes whose keys (values) are less than the key of the current node. In the classic implementation, the binary tree does not store the same keys. In my implementation, this is possible, nodes whose keys are greater or equal to the current one are stored in the left branch.

You can add all values, if they implement Clone, Copy, Eq, Ord traits. It was made to not broke tree's logic (NAN value).

Tree can be used as sorted multiset

std::multiset // C++

from C++: you can collect equal and not equal values in a sorted order, make some manipulations with tree like insertion, clearing, removing values, change equal values to other values and so on. Check method list and examples to be familiar with my crate. Also you can visit the project's repository https://github.com/dinaraparanid/Binary-Tree And check the full code. I highly recommend to see tests.rs file. There are a lot of examples of code. so you'll easily figure out how to use it

Russian: О СТРУКТУРЕ

Бинарное дерево поиска - условно направленный граф, каждый узел которого имеет ключ - некое значение, которое можно сравнивать и две ветви - правую и левую. (будем считать, что ветви есть всегда, просто они могут быть либо пустыми, либо содержать значение).

В левой ветви хранятся узлы, ключи (значения) которых меньше ключа текущего узла. В классической реализации бинарное дерево не хранит одинаковые ключи. В моей реализации это возможно, узлы, ключи которых больше или равны текущему хранятся в левой ветви.

Дерево может хранить только те типы, которые реализуют трейты Clone, Copy, Ord, Eq. Это сделанно с целью не нарушать логику дерева (NAN)

Дерево можно использовать как отсортированный мультисэт

std::multiset // C++

из C++: Можно хранить одинаковые ключи, добавлять или удалять их из дерева, заменять одинаковые ключи на другие значения. Изучите список методов и примеры для полного понимания возможностей дерева. Так же советую посмотреть репозиторий проекта https://github.com/dinaraparanid/Binary-Tree на котором есть весь код. Особенно стоит изучить tests.rs файл. Там много примеров того, как всё это можно использовать.

English: The tree itself. Contains the top node (start of tree) and it's size.

Russian: Само дерево. Храним головной узел - начало дерева, и размер.

Implementations

impl<T> BinaryTree<T> where
T: Copy + Clone + Ord + Eq,

Realisation of tree's methods.

pub fn new() -> Self

English: Method new() creates empty tree. You can call it a constructor

Russian: Метод new() создаёт пустое дерево. По сути является конструктором для дерева.

Example

use binartree::tree::BinaryTree;

let new_tree = BinaryTree::<i32>::new();

assert_eq!(new_tree.len(), 0);
assert_eq!(new_tree.to_vec(), vec![]);

pub fn len(&self) -> usize

English: Method len() returns tree's length.

Russian: Метод len() возвращает длину дерева.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();
assert_eq!(tree.len(), 0);

tree.insert(&1);
tree.insert(&2);
tree.insert(&3);

assert_eq!(tree.len(), 3);

pub fn is_empty(&self) -> bool

English: Method is_empty() answers the question: "is out tree empty?" If it's true, returns true, else false.

Rusiian: Метод is_empty() отвечает на вопрос: "пусто ли наше дерево?" Если пусто, то возвращает true, иначе false.

Example

use binartree::tree::BinaryTree;

let empty_tree = BinaryTree::<i32>::new();
let mut not_empty_tree = BinaryTree::new();

not_empty_tree.insert(&1);
not_empty_tree.insert(&2);
not_empty_tree.insert(&3);

assert_eq!(empty_tree.len(), 0);
assert_eq!(empty_tree.is_empty(), true);

assert_eq!(not_empty_tree.len(), 3);
assert_eq!(not_empty_tree.is_empty(), false);

pub fn to_vec(&self) -> Vec<T>

English: Method to_vec() converts tree to std::vec::Vec

Russian: Метод to_vec() конвертирует дерево в std::vec::Vec

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let tree = BinaryTree::from_iter((1..6));
assert_eq!(tree.to_vec(), vec![1, 2, 3, 4, 5]);

pub fn to_deque(&self) -> VecDeque<T>

English: Method to_deque() converts tree to std::collections::VecDeque

Russian: Метод to_deque() конвертирует дерево в std::collections::VecDeque

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;
use std::collections::VecDeque;

let tree = BinaryTree::from_iter((1..6));
assert_eq!(tree.to_deque(), VecDeque::from_iter(1..6));

pub fn insert(&mut self, val: &T)

English: Method insert() adds value to tree. Value takes by immutable reference (&T), so there is no ownership.

Russian: Метод insert() добавляет значение в дерево. Значение берётся по неизменяемой ссылке (&T), так что владение исключено.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();
assert_eq!(tree.len(), 0);

tree.insert(&1);
tree.insert(&2);
tree.insert(&3);

assert_eq!(tree.len(), 3);
assert_eq!(tree.contains(&1), true);
assert_eq!(tree.contains(&2), true);
assert_eq!(tree.contains(&3), true);
assert_eq!(tree.to_vec(), vec![1, 2, 3]);

pub fn contains(&self, val: &T) -> bool

English: Method contains() checks that value is in the tree. If it's true, returns true, else false

Russian: Метод contains() проверяет наличие значения в дереве Если оно есть, то возвращает true, иначе false

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();
tree.insert(&1);
tree.insert(&3);

assert_eq!(tree.len(), 2);
assert_eq!(tree.contains(&1), true);
assert_eq!(tree.contains(&2), false);
assert_eq!(tree.contains(&3), true);
assert_eq!(tree.to_vec(), vec![1, 3]);

pub fn first(&self) -> &T

English: Method first() returns minimum value in the tree. if tree is empty, it'll panic.

Russian: Метод first() возвращает минимальное значение в дереве. Если дерево пустое, то будет вызвана паника.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();

tree.insert(&3);
assert_eq!(tree.first(), &3);

tree.insert(&4);
 assert_eq!(tree.first(), &3);

tree.insert(&1);
 assert_eq!(tree.first(), &1);

pub fn last(&self) -> &T

English: Method last() returns maximum value in the tree. if tree is empty, than it'll panic.

Russian: Метод first() возвращает максимальное значение в дереве. Если дерево пустое, то будет вызвана паника.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();

tree.insert(&1);
assert_eq!(tree.last(), &1);

tree.insert(&0);
assert_eq!(tree.last(), &1);

tree.insert(&3);
assert_eq!(tree.last(), &3);

pub fn iter(&self) -> TreeIter<T>

English: Method iter() converts tree to iterator, which collects elements in sorted order, It takes tree by immutable reference, so it'll be no ownership. Check iter.rs for TreeIter.

Russian: Метод iter() превращает дерево в итератор, который хранит элементы по-возрастанию. Берётся неизменяемая ссылка, так что владения нет. Изучите iter.rs для полного понимания.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();
tree.insert(&1);
tree.insert(&2);
tree.insert(&3);

assert_eq!(tree.iter().collect::<Vec<i32>>(), vec![1, 2, 3]);

pub fn append(&mut self, src: &BinaryTree<T>)

English: Method apend() translates all elements from 2-nd tree to 1-st. All trees are taking by immutable reference, no ownership.

Russian: Метод append() передаёт элементы из 2 дерева в 1. Используемое дерево остаётся неизменным, так что его можно использовать повторно.

Example

use binartree::tree::BinaryTree;

let mut tree1 = BinaryTree::new();
let mut tree2 = BinaryTree::new();

tree1.insert(&1);
tree2.insert(&2);
tree2.insert(&3);

tree1.append(&tree2);

assert_eq!(tree1.to_vec(), vec![1, 2, 3]);

pub fn clear(&mut self)

English: Method celan() makes tree empty.

Russian: Метод clean() полностью очищает дерево.

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let mut tree = BinaryTree::from_iter((1..11));
assert_eq!(tree.to_vec(), vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);

tree.clear();
assert_eq!(tree.to_vec(), vec![]);

pub fn remove(&mut self, val: &T)

English: Removing element from tree. Keys, that are in subnodes are still in tree. USE THIS METHOD ONLY IF YOU WANT TO REMOVE ONE ELEMENT FROM TREE. USE MULTI_REMOVE() IF YOU WISH TO REMOVE MORE THAN 1 ELEM. IT'S A LOT MORE FASTER

Russian: Удаление элемента из дерева. Ключи, которые есть в подузлах искомого узла остаются в нашем дереве. ИСПОЛЬЗУЙТЕ ЭТОД МЕТОД ТОЛЬК ЕСЛИ ХОТИТЕ УДАЛИТЬ 1 ЭЛЕМЕНТ ИЗ ДЕРЕВА. ЛУЧШЕ ИСПОЛЬЗОВАТЬ MULTI_ERMOVE(), КОТОРЫЙ РАБОТАЕТ В РАЗЫ БЫСТРЕЕ

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let mut tree = BinaryTree::from_iter((1..11));
assert_eq!(tree.to_vec(), vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);

tree.remove(&5);
tree.remove(&1);
assert_eq!(tree.to_vec(), vec![2, 3, 4, 6, 7, 8, 9, 10]);

pub fn difference(&self, other: &BinaryTree<T>) -> TreeIter<T>

English: Method difference() returns TreeIter, which contains all elements, that are in 1-st tree, but not in 2-nd,

Russian: Метод difference() возвращает "деревянный итератор", который хранит все те элементы, которые есть в 1 дереве, но которых нет во 2.

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let tree_1 = BinaryTree::from_iter((1..4));
let tree_2 = BinaryTree::from_iter((3..4));

assert_eq!(tree_1.difference(&tree_2).collect::<Vec<i32>>(), vec![1, 2]);

pub fn drain_filter<F: FnMut(&T) -> bool>(&mut self, fun: F) -> TreeIter<T>

English: Method drain_filter() stoles all values from tree, which are match the simple or lambda function. Returns iterator with all removed values.

Russian: Метод drain_filter() крадёт все элементы из дерева, которые соответстуют той функции или замыканию. Возвращает итератор с удалёнными значениями.

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let mut tree1 = BinaryTree::from_iter(1..11);
let tree2 = tree1.drain_filter(|x| x % 2 == 0);

assert_eq!(tree1.iter().collect::<Vec<i32>>(), vec![1, 3, 5, 7, 9]);
assert_eq!(tree2.collect::<Vec<i32>>(), vec![2, 4, 6, 8, 10]);

pub fn intersection(&self, other: &BinaryTree<T>) -> TreeIter<T>

English: Method intersection() returns TreeIter, which contains all elements, that are in 1-st and 2-nd tree

Russian: Метод intersection() возвращает "деревянный итератор", который хранит все те элементы, которые есть и в 1, и во 2 дереве

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let tree_1 = BinaryTree::from_iter((1..10));
let tree_2 = BinaryTree::from_iter((1..15));

assert_eq!(tree_1.intersection(&tree_2).collect::<Vec<i32>>(), vec![1, 2, 3, 4, 5, 6, 7, 8, 9]);

pub fn is_disjoint(&self, other: &BinaryTree<T>) -> bool

English: Method is_disjoint() answers the question "Are any elements in common?" If yes than false else true

Russian: Метод is_disjoint() отвечает на вопрос "есть ли у деревьев хотя бы 1 общий элемент?" Если да, то false, иначе true

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let tree_1 = BinaryTree::from_iter((1..10));
let tree_2 = BinaryTree::from_iter((10..15));
assert_eq!(tree_1.is_disjoint(&tree_2), true);

let tree_3 = BinaryTree::from_iter((1..10));
let tree_4 = BinaryTree::from_iter((1..15));
assert_eq!(tree_3.is_disjoint(&tree_4), false);

pub fn pop_first(&mut self)

English: Method pop_frist removes min value from tree. If tree is empty, it'll panic

Russian: Метод pop_first() удаляет наименьший элемент дерева. Т.к. дерево может быть пустым, то будет вызвана паника

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;
use std::convert::TryInto;

let mut tree = BinaryTree::from_iter((1..6));
tree.pop_first();
tree.pop_first();
assert_eq!(tree.to_vec(), vec![3, 4, 5]);

pub fn pop_last(&mut self)

English: Method pop_last() removes min element from tree. If tree is empty, it'll panic.

Russian: Метод pop_last() удаляет наибольший элемент дерева. Т.к. дерево может быть пустым, то вызывается паника

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;
use std::convert::TryInto;

let mut tree = BinaryTree::from_iter((1..6));

tree.pop_last();
tree.pop_last();
assert_eq!(tree.to_vec(), vec![1, 2, 3]);

pub fn replace_val(&mut self, old_val: &T, new_val: &T)

English: Method replace_val() changes all keys with some value to another value.

Russian: Метод replace_val() заменяет все ключи с одним значением на ключи с другим значением.

Example

use binartree::tree::BinaryTree;

let mut tree = BinaryTree::new();
tree.extend(vec![1, 1, 1]);

tree.replace_val(&2, &2);
assert_eq!(tree.to_vec(), vec![1, 1, 1]);

tree.replace_val(&1, &3);
assert_eq!(tree.to_vec(), vec![3, 3, 3]);

pub fn symmetric_difference(&self, other: &BinaryTree<T>) -> TreeIter<T>

English: Method symmetric_difference() returns TreeIter with keys, that are only in 1-st or 2-nd tree

Russian: Метод symmetric_difference() возвращает "деревянный" итератор с ключами, которые есть либо только в 1 дереве, либо только во 2.

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let tree1 = BinaryTree::from_iter((1..11));
let tree2 = BinaryTree::from_iter((5..16));

assert_eq!(tree1.symmetric_difference(&tree2).collect::<Vec<i32>>(), vec![1, 2, 3, 4, 11, 12, 13, 14, 15]);

pub fn multi_remove(&mut self, src: Vec<T>)

English: You should use method multi_remove() when you want to remove more than one value from tree. It's a lot more faster then removing elem-by-elem. It takes ownership, so you need to clone it, if you want to use src twice.

Russian: Метод multi_remove() нужен для быстрого удаления сразу нескольких элементов. Работает ГОРАЗДО быстрее, чем поэлементное удаление. Метод принимает владение значениями, так что нужно копировать вектор с ресурсами, если хотите использовать его дважды.

Example

use binartree::tree::BinaryTree;
use std::iter::FromIterator;

let mut tree = BinaryTree::from_iter(1..10);
assert_eq!(tree.to_vec(), vec![1, 2, 3, 4, 5, 6, 7, 8, 9]);

tree.multi_remove(vec![1, 2, 3, 4, 5]);
assert_eq!(tree.to_vec(), vec![6, 7, 8, 9]);